+++ rap/world_model.py
